chore: reorg code; move logic to Method.tsx
it no longer makes sense to keep this logic in `Methods.tsx`

performance *might* be worse now – it feels slightly smaller, but the
Performance tab in the web inspector is inconclusive at first
investigation.

Even if perf is slightly worse with this approach, the code structure
makes more sense, and we can find ways to squeeze out more performance
later.