fix: improve a11y; no overflow scroll in firefox
The overflow scrolling in Firefox was caused by the hidden menu, which
should have only been rendered when in mobile (and now is), but was
rendered all the time.

Getting this to work properly with a screenreader was harder than
expected! It now works the way I would expect: if the menu is visually
hidden, it is hidden from screenreaders as well (the only way to
consistently do this was to use `display: none`).

This also fixes some other accessibility (a11y) issues:

* prevent screenreaders from saying "times sign" for the close button
* after navigating to a new method, get it to read the main content
  (which just changed) right away
* put the sidebar headers into `label` tags, so that clicking anywhere
  on the heading expands/collapses the section, rather than needing to
  click precisely on the little chevron icon
* add more context to screenreader-only button text