feat: contract-specific authentication
Rather than returning `wallet` from `near#init`, return `currentUser`.
This needs to make an RPC query to figure out if the key that NAJ stored
in localStorage is for the current contract or another, making
`currentUser` a promise (the rest of `init` is synchronous, and much of
the codebase relies on it staying that way).