improve some names, to see more clearly
couldn't remember wth "ContractForm" was and how it was different from
"Contract" and "Form". "ContractNameForm" is more accurate--it's the
little form where you enter the name of a contract!

this renames "Contract" to "NearContract" to differentiate it from the
upcoming "CWContract", and nests "Form" within it, since Form was
already a child component in reality but not in file structure.