get some CosmWasm methods listing in sidebar
these are probably incorrect still, but it took a lot of reworking of
the entire codebase to even get it set up for this, so it's time to
commit!

getting valid method names out of the JSONSchema returned by CosmWasm is
not as straightforward as getting it out of the RAEN-generated
JSONSchema.